Take bitcore-lib version from package.json
When using custom versions of bitcore-lib, be in testing or custom
setups, the version that must be installed when creating a node shall
match the one in package.json and not necessarily the one in
`bitcore.version` property.

So this change allows using any custom version of `bitcore-lib` without
having to modify the code.