Add option to specify a custom bitcoind binary
When using custom builds of bitcoind, the path to the binary has to be
configured in `bitcore-node.json`. It is much easier if the path can be
set as a parameter during node creation as it is done with the network.

With this change, the new option looks like:

```
bitcore-node create --exec <full-path-to-custom-bitcoind> mynode
```