Add option to start command to set network params
Definitions for custom networks could be already added to custom
versions of `bitcore-lib` or can be set on start using this new
command-line option:

```
$ bitcore-node --network ./customnet.json
```

The config file will have a JSON that will be sent to
`bitcore.Networks.add()`.

In addition, before starting the services, a check is added to ensure
the custom network exists in the case such a network is being used.